Modified Seats/AvailableSeats/RequestsSeats (Any seat related filed) to be short or smallint(database)

+++ Vehicle_Share/Vehicle_Share.Core/Models/TripModels/GetTripDriverModel.cs
@@ -7,7 +7,7 @@
         public string? To { get; set; }
         public DateTime Date { get; set; }
         public float RecommendPrice { get; set; }
-        public int AvailableSeats { get; set; } //driver
+        public short AvailableSeats { get; set; } //driver
 
         public string? CarId { get; set; } //driver
         public string? CarType { get; set; } //driver
