+++ Vehicle_Share/Vehicle_Share.EF/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -269,8 +269,8 @@ namespace Vehicle_Share.EF.Migrations
                     b.Property<string>("Id")
                         .HasColumnType("nvarchar(450)");
 
-                    b.Property<int?>("AvailableSeats")
-                        .HasColumnType("int");
+                    b.Property<short?>("AvailableSeats")
+                        .HasColumnType("smallint");
 
                     b.Property<string>("CarId")
                         .IsRequired()
@@ -287,8 +287,8 @@ namespace Vehicle_Share.EF.Migrations
                     b.Property<float>("RecommendPrice")
                         .HasColumnType("real");
 
-                    b.Property<int?>("RequestedSeats")
-                        .HasColumnType("int");
+                    b.Property<short?>("RequestedSeats")
+                        .HasColumnType("smallint");
 
                     b.Property<string>("To")
                         .IsRequired()
